Move release-related code to their own module
