+++ xtask/README.md
@@ -10,12 +10,11 @@ Usage: xtask <COMMAND>
 Commands:
   build            Build-related subcommands
   run              Run-related subcommands
-  bump-version     Bump the version of the specified package(s)
+  release          Release-related subcommands
   ci               Perform (parts of) the checks done in CI
   fmt-packages     Format all packages in the workspace with rustfmt
   lint-packages    Lint all packages in the workspace with clippy
-  publish          Attempt to publish the specified package
-  tag-releases     Generate git tags for all new package releases
+  semver-check     Semver Checks
   check-changelog  Check the changelog for packages
   help             Print this message or the help of the given subcommand(s)
